Move Converter to separate file
