+++ src/JuliusSweetland.OptiKey/Models/KeyValue.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Configuration;
-using System.Globalization;
 using System.Runtime.Serialization;
 using System.Security.Permissions;
 using System.Text;
@@ -154,43 +153,6 @@ namespace JuliusSweetland.OptiKey.Models
         public virtual bool HasContent() {            
             return (FunctionKey != null) ||
                    (String != null); 
-        }
-    }
-
-    public sealed class KeyValueConverter : TypeConverter
-    {
-        // The ITypeDescriptorContext interface provides the context for the
-        // conversion. Typically, this interface is used at design time to 
-        // provide information about the design-time container.
-        public override bool CanConvertFrom(ITypeDescriptorContext context,
-           Type sourceType)
-        {
-            if (sourceType == typeof(string))
-            {
-                return true;
-            }
-            return base.CanConvertFrom(context, sourceType);
-        }
-
-        public override object ConvertFrom(ITypeDescriptorContext context,
-           CultureInfo culture, object value)
-        {
-            var text = value as string;
-            if (text!=null)
-            {
-                return new KeyValue(text);
-            }
-            return base.ConvertFrom(context, culture, value);
-        }
-
-        public override object ConvertTo(ITypeDescriptorContext context,
-           CultureInfo culture, object value, Type destinationType)
-        {
-            if (destinationType == typeof(string))
-            {
-                return ((KeyValue)value).String;
-            }
-            return base.ConvertTo(context, culture, value, destinationType);
         }
     }
 }
